Run at startup

+++ mdm/common/src/main/java/com/gentics/mesh/core/jobs/ImageCacheMigrationProcessor.java
@@ -48,11 +48,11 @@ public class ImageCacheMigrationProcessor implements SingleJobProcessor {
 					.filter(path -> path.getFileName().toString().startsWith("image-") && Files.isRegularFile(path))
 					.map(path -> {
 						String sha512Hash = imageCachePath.relativize(path.getParent()).toString().replace("/", "").replace("\\", "");
-						if (sha512Hash.length() == 128) {
-							HibBinary binary = dao.findByHash(sha512Hash).runInExistingTx(Tx.get());
+						HibBinary binary = null;
+						if (sha512Hash.length() == 128 && (binary = dao.findByHash(sha512Hash).runInExistingTx(Tx.get())) != null) {
 							String uuid = binary.getUuid();
 							String segments = getSegmentedPath(uuid);
-							Path segmentsPath = imageCachePath.resolve(segments);
+							Path segmentsPath = Path.of(options.getImageOptions().getImageCacheDirectory(), segments);
 							try {
 								Files.createDirectories(segmentsPath);
 								Files.copy(path, segmentsPath.resolve(uuid + "-" + path.getFileName().toString().replace("image-", "")));
@@ -60,7 +60,7 @@ public class ImageCacheMigrationProcessor implements SingleJobProcessor {
 								log.error("Could not copy old cached file " + path, e);
 							}
 						} else {
-							log.error("Not a SHA512 hash: " + sha512Hash);
+							log.error("Not a SHA512 hash or binary not found: " + sha512Hash);
 						}
 						return path;
 					}).forEach(path -> {
